web and single worker

+++ cloud-config/concourse-web.tpl
@@ -8,15 +8,21 @@ if ! which concourse; then
   mv concourse /usr/local/bin/concourse
 fi
 
-mkdir -p keys/web keys/worker
+aws s3 cp s3://${keys_bucket}/ ./keys --recursive
 
-ssh-keygen -t rsa -f ./keys/web/tsa_host_key -N ''
-ssh-keygen -t rsa -f ./keys/web/session_signing_key -N ''
+if [ ! -f ./keys/web/tsa_host_key ]; then
+    mkdir -p keys/web keys/worker
 
-ssh-keygen -t rsa -f ./keys/worker/worker_key -N ''
+    ssh-keygen -t rsa -f ./keys/web/tsa_host_key -N ''
+    ssh-keygen -t rsa -f ./keys/web/session_signing_key -N ''
 
-cp ./keys/worker/worker_key.pub ./keys/web/authorized_worker_keys
-cp ./keys/web/tsa_host_key.pub ./keys/worker
+    ssh-keygen -t rsa -f ./keys/worker/worker_key -N ''
+
+    cp ./keys/worker/worker_key.pub ./keys/web/authorized_worker_keys
+    cp ./keys/web/tsa_host_key.pub ./keys/worker
+
+    aws s3 cp ./keys s3://${keys_bucket}/ --recursive
+fi
 
 touch /var/log/concourse_web.log
 chmod 666 /var/log/concourse_web.log
